Intermediate commit: trying things with an async generator, but doesn't work because the promise can be overwritten #3

+++ tests/lib/db.js
@@ -1,13 +1,23 @@
 const knex = require('knex')
 const crypto = require('crypto')
 
+// Can be useful if you're doing a single test and want to observe the data after the test has run
+const forceSingleDb = true
+
 async function createRandomTestDb() {
-  const testDbName = 'test_' + crypto.randomBytes(5).toString('hex')
+  let testDbName
+  testDbName = 'test_' + crypto.randomBytes(5).toString('hex')
+  if (forceSingleDb) {
+    testDbName = 'test_db'
+  }
 
   db = knex({
     client: 'mysql2',
     connection: `mysql://${process.env.TEST_DB_ADMIN_USER}:${process.env.TEST_DB_ADMIN_PASSWORD}@localhost:33033/`
   })
+  if (forceSingleDb) {
+    await db.raw(`DROP DATABASE IF EXISTS ${testDbName}`)
+  }
   await db.raw(`CREATE DATABASE ${testDbName}`)
   await db.raw(`GRANT ALL PRIVILEGES ON ${testDbName}.* TO user1`)
   db.destroy()
@@ -19,6 +29,9 @@ async function createRandomTestDb() {
 }
 
 async function dropAndDestroyTestDb(db) {
+  if (forceSingleDb) {
+    return
+  }
   const testDbName = db.client.database()
   await db.raw(`DROP DATABASE ${testDbName}`)
   db.destroy()
